refactor: `TxProcessor` to support pluggable indices
